feat: добавить публичный поиск Events
