+++ main-service/src/main/java/ru/practicum/mainservice/category/dto/CategoryDto.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class CategoryDto {
-    private Long Id;
+    private Long id;
     private String name;
 }
