+++ main-service/src/main/java/ru/practicum/mainservice/category/model/Category.java
@@ -1,8 +1,6 @@
 package ru.practicum.mainservice.category.model;
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
