+++ main-service/src/main/resources/application.properties
@@ -1,5 +1,5 @@
 server.port=8080
-#spring.mvc.format.date-time=yyyy-MM-dd HH:mm:ss
+spring.mvc.format.date-time=yyyy-MM-dd HH:mm:ss
 
 spring.jpa.hibernate.ddl-auto=none
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQL10Dialect
