+++ main-service/src/main/resources/schema.sql
@@ -25,6 +25,7 @@ CREATE TABLE IF NOT EXISTS events
     event_id BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
     annotation  VARCHAR(2000)    NOT NULL,
     created TIMESTAMP WITHOUT TIME ZONE NOT NULL,
+    published TIMESTAMP WITHOUT TIME ZONE,
     category_id    BIGINT REFERENCES categories ON DELETE CASCADE,
     description  VARCHAR(7000)    NOT NULL,
     event_date TIMESTAMP WITHOUT TIME ZONE NOT NULL,
@@ -33,6 +34,7 @@ CREATE TABLE IF NOT EXISTS events
     location_lon FLOAT NOT NULL,
     paid BOOLEAN NOT NULL,
     participant_limit INTEGER NOT NULL,
+    confirmed_requests INTEGER NOT NULL,
     request_moderation BOOLEAN NOT NULL,
     state  VARCHAR(30)    NOT NULL,
     title  VARCHAR(120)    NOT NULL,
